Create TicTacToe
My first code with the help with techwithtim at https://www.techwithtim.net/tutorials/python-programming/tic-tac-toe-tutorial/.  Mad some corrections to the code to adjust the lines of the grids.  Also corrected the while loop so that the game can continue if player choses to. I also added a print statement after a player selects "no".  I sitll have a lot of work to do as far as understanding the characters play in the code.